requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astroid==3.0.2
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
dill==0.3.7
diskcache==5.6.3
distro==1.8.0
Django==4.2.6
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==22.2.0
frozenlist==1.4.0
gptcache==0.1.43
guidance==0.1.5
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
inflection==0.5.1
isort==5.13.2
joblib==1.3.2
llama_cpp_python==0.2.19
lxml==4.9.3
mccabe==0.7.0
msal==1.26.0
multidict==6.0.4
mysqlclient==2.2.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.3
openai==1.3.7
ordered-set==4.1.0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
platformdirs==4.1.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
pyformlang==1.0.4
PyJWT==2.8.0
pylint==3.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tiktoken==0.5.2
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
XlsxWriter==3.1.9
yarl==1.9.3
